+++ at_client/lib/src/service/encryption_service.dart
@@ -456,4 +456,18 @@ class EncryptionService {
         EncryptionUtil.decryptKey(encryptedSharedKey, currentAtSignPrivateKey);
     return sharedKey;
   }
+
+  String generateFileEncryptionKey() {
+    return EncryptionUtil.generateAESKey();
+  }
+
+  List<int> encryptFile(List<int> fileContent, String fileEncryptionKey) {
+    return EncryptionUtil.encryptBytes(fileContent, fileEncryptionKey);
+  }
+
+  List<int> decryptFile(List<int> fileContent, String decryptionKey) {
+    var encryptedValue =
+        EncryptionUtil.decryptBytes(fileContent, decryptionKey);
+    return encryptedValue;
+  }
 }