+++ at_client/lib/src/util/constants.dart
@@ -1,0 +1,4 @@
+class TextConstants {
+  static const String FILEBIN_URL = 'https://filebin2.aws.atsign.cloud/';
+  static const String FILE_TRANSFER_KEY = 'file_transfer_';
+}