+++ at_client/test/samples/remote_lookup.dart
@@ -8,7 +8,7 @@ void main() async {
     var preference = TestUtil.getPreferenceRemote();
     await AtClientImpl.createClient('@alice', null, preference);
     var atClient = await (AtClientImpl.getClient('@alice'));
-    if(atClient == null) {
+    if (atClient == null) {
       print('unable to create at client instance');
       return;
     }