+++ at_client/test/samples/remote_plookup.dart
@@ -8,7 +8,7 @@ void main() async {
     await AtClientImpl.createClient(
         '@alice', null, TestUtil.getPreferenceRemote());
     var atClient = await (AtClientImpl.getClient('@alice'));
-    if(atClient == null) {
+    if (atClient == null) {
       print('unable to create at client instance');
       return;
     }