+++ at_client/CHANGELOG.md
@@ -30,6 +30,10 @@
 - Minor bug in stream handlers
 ## 2.0.0
 - Null safety upgrade
+## 2.0.1
+- at_commons version upgrade
+## 2.0.2
+- filebin upload changes
 ## 3.0.0
 - Changes to sync, monitor and notify implementation to handle network failures
 and retries.