+++ at_client_mobile/CHANGELOG.md
@@ -28,5 +28,8 @@
 - Provision to request for a new outbound connection.
 - Minor bug in stream handlers
 - at_client version upgrade
+## 2.0.0
+- Null safety changes
+- Dependent package upgrade
 ## 3.0.0
 - Null safety changes